pyproject.toml
[project]
name = "seqai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb",
    "transit-python2"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ipython"
]

[tool.pytest.ini_options]
pythonpath = ["."]
